requirements.txt
# Core Django Framework
Django==5.2.4

# Environment Management
python-dotenv==1.1.1

# HTTP Requests for API Integration
requests==2.32.4

# Google Gemini AI Integration
google-generativeai==0.8.5
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.177.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
googleapis-common-protos==1.70.0

# Additional Dependencies (automatically installed with above)
# These are included for completeness and version tracking
asgiref>=3.6.0
sqlparse>=0.4.2
tzdata>=2023.3

# Development Dependencies (optional)
# Uncomment for development environment
# flake8==6.0.0
# pytest==7.4.0
# pytest-django==4.5.2
# coverage==7.2.0

# Production Dependencies (optional)
# Uncomment for production deployment
# gunicorn==21.2.0
# psycopg2-binary==2.9.7
# whitenoise==6.5.0
# redis==4.6.0
